requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv
openai>=1.0.0
chromadb
